requirements.txt
imutils
pillow
scikit-image
numpy
paddlepaddle
paddleocr
opencv-python
opencv-python-headless
streamlit
streamlit-paste-button
py-sudoku
picwish